rename: review-component
