+++ s/icons/bin.svg.ts
