+++ s/icons/danger.svg.ts
