+++ s/icons/info-square.svg.ts
