+++ thrills-Test.html
@@ -17,18 +17,18 @@
 			padding: 10vh 0;
 		}
 
-		review-comment {
+		namaka-review-comment {
 			margin: auto;
 		}
 	</style>
 </head>
 <body>
 	<div class="container">
-		<review-comment></review-comment>
-		<review-comment></review-comment>
-		<review-comment></review-comment>
-		<review-comment></review-comment>
-		<review-comment></review-comment>
+		<namaka-review-comment></namaka-review-comment>
+		<namaka-review-comment></namaka-review-comment>
+		<namaka-review-comment></namaka-review-comment>
+		<namaka-review-comment></namaka-review-comment>
+		<namaka-review-comment></namaka-review-comment>
 	</div>
 </body>
 </html>
